Addressed final comment.
